Refactor codebase: Add JSDoc comments, fix imports, and organize types
