+++ src/commands/stack/registry.ts
@@ -8,7 +8,9 @@ import { refreshStackView, refreshActiveStack } from ".";
  * Registers stack-related commands for the extension.
  *
  * @param {vscode.ExtensionContext} context - The context in which the extension operates, used for registering commands and managing their lifecycle.
+ * @param {Shell} shell - An instance of Shell that provides access to the shell environment for executing commands.
  * @param {StackDataProvider} stackDataProvider - An instance of StackDataProvider that manages the data and updates the view for stack-related operations.
+ * @param {ZenMLStatusBar} statusBar - An instance of ZenMLStatusBar that manages the status bar for the extension.
  */
 export function registerStackCommands(
   context: vscode.ExtensionContext,
