+++ src/types/StackTypes.ts
@@ -1,0 +1,10 @@
+/**
+ * Represents a component within a stack in the tree view.
+ */
+export interface StackComponent {
+  id: string;
+  name: string;
+  type: string;
+  flavor: string;
+}
+
